Add MenubarTest to AppHeader story in Storybook
This commit enhances the AppHeader story in Storybook by importing and including MenubarTest component. It also changes the export name from 'Default' to 'MenubarTest' in MenuBar.stories.tsx. Lastly, additional styling properties are introduced in AppHeader.tsx. This addition aims to enrich the presentation of the AppHeader component and its interaction with the Menubar in Storybook.